+++ CHANGELOG.md
@@ -12,6 +12,7 @@ NOTE: Bump file format version to 21. NO DOWNGRADE PATH IS AVAILABLE.
 * Array of primitive lists will not be `snapshot()`'ed.
 * Added `ssl` option to `Realm.App.Sync` configuration.
 * Added UUID types. ([#3244](https://github.com/realm/realm-js/issues/3244))
+* Adding Dictionaries types. ([#3392](https://github.com/realm/realm-js/issues/3392))
 
 ### Fixed
 * <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-js/issues/????), since v?.?.?)
@@ -23,7 +24,7 @@ NOTE: Bump file format version to 21. NO DOWNGRADE PATH IS AVAILABLE.
 * File format: generates Realms with format 21.
 
 ### Internal
-* Bump the Realm Sync protocol version to 3.
+* Bump the Realm Sync protocol version to 4.
 * Bump Realm File Format version to 21.
 * Prebuild the React Native iOS variant and bundle it in the npm tarball. ([#3649](https://github.com/realm/realm-js/pull/3649))
 * Upgraded Realm Core to v11.0.0-beta.3.